ingestion/prefect_agent/service_asr/pyproject.toml
[project]
name = "service-asr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.0",
    "chunkformer>=1.2.1",
    "colorama>=0.4.6",
    "fastapi>=0.118.2",
    "ffmpeg-python>=0.2.0",
    "huggingface-hub>=0.35.3",
    "ipykernel>=6.30.1",
    "jiwer>=4.0.0",
    "jupyter>=1.1.1",
    "loguru>=0.7.3",
    "minio>=7.2.18",
    "moviepy>=2.2.1",
    "numpy<2",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.3",
    "pillow>=11.3.0",
    "pydantic>=2.12.0",
    "pydantic-settings>=2.11.0",
    "pydub>=0.25.1",
    "python-consul2[aio]>=0.1.5",
    "pyyaml>=6.0.3",
    "sentencepiece>=0.2.1",
    "textgrid>=1.6.1",
    "torch>=2.8.0",
    "torchaudio>=2.8.0",
    "tqdm>=4.67.1",
    "uvicorn>=0.37.0",
]
